minor changes; move relative configs to experimental/
