+++ configs/experimental/prolificdreamer-propnet.yaml
@@ -1,4 +1,4 @@
-name: "prolificdreamer-prop-sdf"
+name: "prolificdreamer-propnet"
 tag: "${rmspace:${system.prompt_processor.prompt},_}"
 exp_root_dir: "outputs"
 seed: 0
@@ -8,8 +8,8 @@ data:
   batch_size: [1, 1]
   # 0-4999: 64x64, >=5000: 512x512
   # this drastically reduces VRAM usage as empty space is pruned in early training
-  width: [64, 128]
-  height: [64, 128]
+  width: [64, 512]
+  height: [64, 512]
   resolution_milestones: [5000]
   camera_distance_range: [1.0, 1.5]
   fovy_range: [40, 70]
@@ -23,46 +23,79 @@ data:
 system_type: "prolificdreamer-system"
 system:
   stage: coarse
-  geometry_type: "implicit-sdf"
+  geometry_type: "implicit-volume"
   geometry:
-    radius: 2.0
-    normal_type: finite_difference
-    # progressive eps from Neuralangelo
-    finite_difference_normal_eps: progressive
+    radius: 1.0
+    normal_type: null
 
-    sdf_bias: sphere
-    sdf_bias_params: 0.5
+    density_bias: blob_magic3d
+    density_activation: softplus
+    density_blob_scale: 10.
+    density_blob_std: 0.5
 
-    # coarse to fine hash grid encoding
     pos_encoding_config:
-      otype: ProgressiveBandHashGrid
+      otype: HashGrid
       n_levels: 16
       n_features_per_level: 2
       log2_hashmap_size: 19
       base_resolution: 16
-      per_level_scale: 1.381912879967776 # max resolution 2048
-      start_level: 8 # resolution ~200
-      start_step: 2000
-      update_steps: 500
+      per_level_scale: 1.447269237440378 # max resolution 4096
 
   material_type: no-material
   material:
     n_output_dims: 3
     color_activation: sigmoid
 
-  background_type: "solid-color-background"
+  background_type: neural-environment-map-background
   background:
-    n_output_dims: 3
-    color: [0, 0, 0]
+    color_activation: sigmoid
+    random_aug: true
 
-  renderer_type: neus-volume-renderer
+  renderer_type: nerf-volume-renderer
   renderer:
     radius: ${system.geometry.radius}
-    use_volsdf: true
-    estimator: importance  # should in ['occgrid', 'importance']
-    num_samples_per_ray: 64
-    num_samples_per_ray_proposal: 128
-
+    estimator: proposal # should in ['occgrid', 'proposal']
+    num_samples_per_ray: 32
+    num_samples_per_ray_proposal: 64
+    near_plane: 0.1
+    far_plane: 4.0
+    proposal_network_config:
+      n_input_dims: 3
+      n_output_dims: 1
+      encoding_config:
+        otype: HashGrid
+        n_levels: 5
+        n_features_per_level: 2
+        log2_hashmap_size: 17
+        base_resolution: 16
+        per_level_scale: 1.681792830507429 # max_resolution: 128
+      network_config:
+        otype: VanillaMLP
+        activation: ReLU
+        output_activation: none
+        n_neurons: 64
+        n_hidden_layers: 1
+    prop_optimizer_config:
+      name: Adam
+      args:
+        lr: 1.0e-2
+        eps: 1.0e-15
+        weight_decay: 1.0e-6
+    prop_scheduler_config:
+      name: ChainedScheduler
+      schedulers:
+        - name: LinearLR
+          args:
+            start_factor: 0.01
+            total_iters: 100
+        # -
+        #   name: MultiStepLR
+        #   args:
+        #     milestones:
+        #       - ${idiv:${trainer.max_steps},2}
+        #       - ${idiv:${mul:${trainer.max_steps},3},4}
+        #       - ${idiv:${mul:${trainer.max_steps},9},10}
+        #     gamma: 0.33
 
   prompt_processor_type: "stable-diffusion-prompt-processor"
   prompt_processor:
@@ -89,10 +122,9 @@ system:
     lambda_vsd: 1.
     lambda_lora: 1.
     lambda_orient: 0.
-    lambda_sparsity: 0.
-    lambda_opaque: 0
+    lambda_sparsity: 10.
+    lambda_opaque: [10000, 0.0, 1000.0, 10001]
     lambda_z_variance: 0.
-    lambda_eikonal: 100.
   optimizer:
     name: AdamW
     args:
@@ -101,13 +133,11 @@ system:
     params:
       geometry.encoding:
         lr: 0.01
-      geometry.sdf_network:
+      geometry.density_network:
         lr: 0.001
       geometry.feature_network:
         lr: 0.001
       background:
-        lr: 0.001
-      renderer:
         lr: 0.001
       guidance:
         lr: 0.0001
